rename confetti.py
